separate cube app, keyboard rework

+++ tilesge/include/game/FpsCounter.h
@@ -9,15 +9,15 @@
 
 class FpsCounter {
 public:
-    void registerFps(float deltaTime);
-    float getFps() const;
+    void registerFps(double deltaTime);
+    double getFps() const;
 private:
     static const size_t fpl = 60;
-    float dts[fpl] = {0,};
+    double dts[fpl] = {0,};
     size_t fpsPointer = 0;
     size_t filled = 1;
 
-    float cachedFps = 0.f;
+    double cachedFps = 0.f;
 };
